New options and dmsquery fix for release-shortlog.py
Added option --no-count to suppress commit count
functionality.

Added option --qry <file> to provide an output file
for dms query.

New usage:
release-shortlog.py [--job <job url>] \
[--not <ref>[,ref,...]] [--no-count] \
[--qry <filename>] <old> <new>

Where <old> and <new> are references to manifests.
The reference can either be a version id to use
in a hudson lookup, a Hudson build job url, or
a manifest file.

Change also includes adaptations and fixes for
generating a dms query file with the correct
contents.

The execution of the dmsquery script seemed to
be waiting for input indefinitley under some
circumstanses. Input to the dmsquery process
is now closed.