Release_tool: Generate report for direct commits and rebased commits.
release-info.py script is based on existing script of release-shortlog.py
and did some restruct and some improvement. It's to meet the KPI requirement of
cnbj SWPM. It is used to generate the DMS report and commits reports for direct
commits and rebased commits between two specified snapshot of one project
against the rebase project.
As we're planning to develop a series of scripts related to release, so I
created a new folder of release_tool to store them.

Change-Id: I597fa3e31b3b40833b175159b2fa9d276e2d4080