Python cherry pick script
This script is reading git log of source branch and destination
branch, finds the Fix=DMSxxxxxxx tag (and other necessary info)
and identifies the commits which are not merged to destination
branch yet. Then it checks the DMS tag in DMS server and
generates a .csv file with list of commits which has correct
tag, cherry pick and push the commits to Gerrit for review.

Please read usage/help (-h) to know how to use this script.