Integration of dms tag server
Checking DMS status through web interface is very slow. It is
very fast if we run dms_tag_server.py in windows server and send
request from this script to the server. To achive this run
dms_tag_server.py in windows machine and pass the ip address of
the server into this script with --dms-tag-server <ip> option.