commitcheck.py: rewrite to check the whole commit message
The commit message checker script (commitcheck.py) is rewritten to
check the content of the entire commit message.  Instead of taking
input from the Gerrit query output, it now takes the output from the
git cat-file command.

Two Hudson scripts are added. One will download a patch set from
Gerrit and run the commit message checker on it, and one will
download the patch set and check it for whitespace errors.