commitcheck-hudson: set review score directly instead of by trigger
The Hudson Trigger Plugin has limitations that make it not possible
to use it to set review scores for commit message checking.

If the commit message job is configured to give 0 verified score on
success, it overrides other jobs that are set to give +1.

However if the commit message job is configured to give the "default"
score, it sets +1 verified on success.

Neither of the above situations are acceptable, therefore we need to
implement the scoring in the script itself, rather than using the
inbuilt scoring mechanism in the trigger plugin.