External command timeout added
The external system commands will be killed if not completed when
the timeout expires. This is handled via a timeout argument added
to the execmd function.

The default timeout value is 30sec. As per current implementation,
all the commands are running with the default timeout, except for
repo sync which has 1 hour.