Target branch patterns implemented
The target branch patterns state the regular expresions that must
match the revision key in the target manifest in order to proceed
with cherry picking for the particular project.
These regular expresions are evaluated apart and after SHA1's and
the default value is '(?!refs/tags)', meaning everything except
tags.
To pass different values, the option --target-branch-patterns was
added and should contain the accepted patterns, comma separated.

Additionally, the function is_str_git_sha1 was added to check if
the input string is a valid revision SHA1 and the fuction
loop_match to check if the input string matches one of several
patterns.