'target_origin' member added to the Commit class
'target_origin' member was added to the Commit class to make
available the correct origin to create the topic branch from. The
default value is 'origin/' + target. This is useful especially
when a sha1 of the target revision is to be replaced with a newly
created branch.

Related corrections were introduced in the get_dms_list and
cherry_pick functions.