Added report_change_impact.py
Intended to be triggered by newly uploaded patch sets, this new script
checks which manifest branch(es) will be affected when a change is
eventually merged. If more than one branch will be affected a note
is added to the change to remind the uploader that additional
verification might be necessary. Extensive filtering capabilities
are available to avoid reporting stuff unnecessarily.

Change-Id: I95cdfafd8a59a2cec8fbf1c0899bd85db611127c