cherrypick.py: Ensure trailing newline in updated manifest
When updating the manifest and writing it back to disk,
minidom's writexml() didn't finish off the XML markup
with a newline character. In most cases this caused an
extra diff in the resulting commit. Not only that, this
missing newline character would probably be added back by
the next human making a commit since many text editors
ensure files end with a newline.

This is fixed by instead using toxml() and adding a
newline character if it isn't already there (in case
toxml() in future versions of Python or libxml2 would
change their behavior).