cherrypick.py: Updated the script to handle new DMS structure
With the introduction of new DMS structure, it is necessary to
pass the target branch name to the dms_tag_server for tag checking.

The change now includes the target branch name value in the call
to dms_tag_server.

This change depends on 2 other changes in run_query.pl and
dms_tag_server.py scripts.