Add find_reviewers.py
This script can be used to find appropriate reviewers for a change,
either by change ID, or by project name and branch name.

Additionally, the query method in the Gerrit class is changed to
include the --current-patch-set option in the query, so that the
results from the query contain the currentPatchSet member.