Initial version of snapshot.py
The snapshot script can be used to create
new C2D snapshots based on earlier snapshots
and/or lists or package files listing packages
to add/replace or remove from the copied
snapshot.
The reason we want this is to be able to
create new snapshots based on approved and
tested official builds, where we add or
replace existing applications with new,
decoupled applications delivered directly
to the C2D repository.

The --copy option accepts the label name
of an existing C2D snapshot and when
given, will add all those packages to
the new snapshot.

The --add and --remove options can be
listed multiple times and accept either
strings or package files.
The --add options requires name/version
pairs to be entered as '<name> <version>'
separated by white space.

The --remove option is mostly useful if
you have copied a snapshot but want to
remove packages from it.

Change-Id: Ic0e2d48c01e069edf72d0ec9b9de487f931ad9ca