Jenkins verification: artifacts not archived when no Python files
When there are no Python files in the change, archiving of the
artifacts was failing due to missing pep8_log.txt

Create a new folder called "out" in the workspace, and store the
pep8_log.txt file in there.  Then the Jenkins job can be configured
to archive out/** as artifacts.