Correcting the method "add_from_file" and adding destructor.
The method add_from_file now does not crash as before. It now also
supports the tag hierarchy incl package-groups in the xml files.
The constructor of the Snapshot class now also creates a temporary
directory for the downloaded debian packages. The directory is
removed in the destructor of the class.

Change-Id: I0e45fd4c0c4d83c5315247066cb7a3f5f1671177