commit_message_check.py: refactoring to improve unit testability
The CommitMessageCheck class is refactored to make it easier use as
a utility class, and to be more unit testable.

New method reset() is added, which will clear all currently found
errors.

check_subject() is modified to also check for utf-8 characters, and
to report error when the subject is too long AND split across lines.

Unit tests are added to test all functions.