run_query.pl: Set the official label irrespective of the tag status
If the branch expected valid tag and if the issues did not have one
those issues were skipped during the update.  This behavior is
changed and all the integrated issues will get the official label.

Appropriate comments will be added to the Notes for issues with
incorrect/untagged fix_for value.