snapshot.py: Reversed order of --dir and --add option parsing
Before, the --add was handled before --dir, which means
that packages added by xml files in dir were overwriting
what was added on the command line. This is usually not
the desired behaviour.
Now --dir is handled before --add and add/replace from
the command line will overwrite what is handled with
files in dir.

Change-Id: If3e9851a6f34099967c0bf7fafaa28257d8018fe