run_query.pl: Bug fix for skipping issues while creating new DRs
If something goes wrong in creating a new DR for a DMS issue, the
logic should be to log the error message and proceed with next
issue in the list.  Instead, it was getting out the method that
creates the new DRs.

Fixed this and added more log message to see where it is actually
failing.

Change-Id: I3f7f29210bdc85def8456a968ac75e883e57e45f