retry.py: add an additional test case for unexpected exception
Add a test case to verify the expected behaviour when the called method
fails, but does not raise the type of exception that is expected in the
decorator on the method's declaration.

Also changed "retries" to "tries" in the test case variable names, as
this more accurately describes the behaviour.