getmanifest.py: Added new interface to extract the control field values
Given any package name and version it is possible to extract the
control field values through the new interface.

By default it works for the 'build-metadata' package if package name
is not given.

Also now the `get_and_extract_package()` method returns the full path
of the downloaded debian package.  This is used by the new method,
`get_control_fields()`.

Updated the script to use the 'debian_bundle' module to handle the
debian packages.

Change-Id: I6bccad2a2d489db68f29e1548f145cd2731e44ce