Cherrypick: Reduced cilent server transaction for dms query
Currently the cherry pick script queries the DMS Tag Server
multiple times for the same set of DMS issues for different
tags.  This increases the network transaction and time for
each run.

Now all the tags are sent in one request. The server side
is updated to check the DMS issues against all tags in one
call and returns a list with all tagged DMS issues to the client.

The dmsutil.py was returning a hash of DMS issues with tag
as the key to the hash.  Now changed to return a list instead.

The cherrypick.py is modified to handle the returned list
instead of hash table.