dms_tag_server.py: Don't check anything if the tag list is empty
Currently the script checks the fix_for field against the received
tag(s) and if the tag list is empty, the DMS in question is not
returned back to the client (this shouldn't be the case when there
is no tag enforced by the system branch).

By not returning the DMS issue the cherry pick script will skip the
related commit even if it is a free bug fix phase.

Or in other words, if the received tag list is empty, all the received
DMS issues should be returned back to the client without checking
anything.