report_change_impact.py: Tidy up the bullet lists
The templates used for producing the DMS violation texts contain
newline characters. This creates a bit of a problem since Gerrit
causes newline characters within a bullet list paragraph to start
new bullet list items. Hence, a text meant to result in a single
bullet will instead be split into multiple bullets. This is solved
by replacing the newlines with spaces when we compose the final
copy of the message.

Change-Id: I56877d2b3da1ae6ca19746d2ec3b262a44611ed0