cherrypick.py: use a unique topic branch name
Sometimes the cherry pick script fails before it has deleted the topic
branch.  This causes an error on the next run when it tries to create
another topic branch with the same name.

To prevent this, create the topic branch with a unique name each time.

The topic branch name is generated based on the sha1 of the commit that
is being cherry picked, and current time.