Add test case to check for spurious executable file mode bits
It seems people are mistakenly setting the executable bit
of files where it doesn't make sense to have the bit set
(text files, XML files, ...) and this test case makes sure
we catch those. For lack of better methods, we use file(1)
to try to figure out which files are executable and hence
exempt from the executable bit blacklist.

Change-Id: I6edbb5daa4ab87a741b7a2912405bf3ea7be1fa4