report_change_impact.py: add support for code review and verify scoring
Updated the script to add code review and verify scores to the Gerrit
review, according to values in the policy configuration, when a change
does not conform to the DMS tagging policy.

Updated the policy for current release branches to add a code review
score of -1.

Also made a couple of tweaks in the informational logging.