report_change_impact.py: check if change is open before scoring
It is possible that a change can be abandoned or merged during the
time that it takes for the script to be invoked and complete.  If this
happens, an error will be raised when trying to add code review and
verify scores.

The script is updated to check if the change is still open, before
attempting to add code review and verify scores.

A new helper function is_change_open is added in the gerrit module.