commit_message_check.py: move functionality from wrapper to script
The script is refactored to move functionality out of the wrapper
shell script into the python script itself.  Parameters are passed
into the script, which then downloads the change and extracts the
commit message.

Also changed to include the error/warning log in the review message,
and only give a code review score if the change is still open and
the patch set is current.