report_change_impact.py: check if patch set is current before scoring
If a new patch set is uploaded during the time it takes for the script
to complete, Gerrit will fail when trying to add code review or verify
scores on the old patch set.

Added a check to prevent code review or verify score from being added
to a patch set that has been replaced with a newer one.