gerrit.py: add get_patchset_refspec and move methods to module scope
Added a new method to get the refspec of a patchset when given the
change number and patchset number.

Also moved the class methods that do not have dependency to the class
instance out of the class scope into module scope, so that they may be
used without having to instantiate the class.