branch_policies.py: Print DMS tag values for a given branch
Currently the 'cm_check_policies' job validates the DMS using the
dms_policy.xml file but the official post build job and DMS update
job use a static value.  Due to this manual update of the official
jobs are required whenever the DMS policy changes.

This change will make the script as a stand alone executable to print
the tag values. The same can be used as a parameter for run_query.pl
during DMS update.

Change-Id: I40542f84bab357e8547520b1aeff452f5ccaa13b