commit_message_check: split functionality in Jenkins wrapper script
The Jenkins wrapper script is split into two files.  The wrapper script
that runs in Jenkins simply downloads the cm_tools git and runs a
second script, inside cm_tools, which actually invokes the commit
message checker.

With this implementation, it is not necessary to update the Jenkins job
every time a modification needs to be made in the configuration.  The
wrapper script is copy/pasted into the Jenkins configuration once, and
then any changes in how the commit_message_checker is invoked can be
made in the second script and merged in the cm_tools git.  When the
change is merged it will automatically be used in the Jenkins job.