commit_message_check: make the Jenkins script more robust
The Jenkins script is updated to be more resilient to errors when
the call to curl (to get the list of PLD gits) fails.  Instead of
causing the whole script to fail, an error on curl will simply
cause the check for PLD gits to be skipped.  As a side-effect of
this, if curl fails, the commit message check will be run on all
gits including PLD gits.

Also updated the check to find the exact git name in the list of
PLD gits.