run_query.pl: Adding '-title' option
This change is to get rid of the dependency to delivery web tool for
generating DMS query file.

The delivery web tool was used to get the DMS title information and
to generate the DMS query file during the official build.  Now the
available android-cm-web will be used for these tasks instead.

As the first step the 'run_query.pl' is updated with one more option,
'-title' to display the DMS title information along with the '-list'
option.

Change-Id: I97b73544675a0164420afb0215946d55c8ffd8e5