dms_tag_server.py: Adding new request type, 'DMS_INFO'
This change is to get rid of the dependency to delivery web tool for
generating DMS query file.

The delivery web tool was used to get the DMS title information and
to generate the DMS query file during the official build.  Now the
available android-cm-web will be used for these tasks instead.

The 'dms_tag_server.py' is updated to respond to a new request,
DMS_INFO.  It will contact the DMS server and retrieve the DMS title
information.

Change-Id: Ib191091a339b6a5fe87edb60e9e15ac60bb03b0f