cm_check_policies: split functionality in Jenkins wrapper script
The Jenkins wrapper script it split into two files.  The wrapper script
that runs inside Jenkins only downloads the cm_tools and manifest gits
and then runs a second new script that exists in cm_tools to actually
invoke the policies check python script.

With this implementation, it is not necessary to update the Jenkins job
every time a modification needs to be made in the configuration and/or
in how the python script is invoked.  The change can be made in the
second script in cm_tools and then when it is merged it it will be
automatically used in the Jenkins job.