dmsutil.py: Added new method to get the DMS title
This change is to get rid of the dependency to delivery web tool for
generating DMS query file.

The delivery web tool was used to get the DMS title information and
to generate the DMS query file during the official build.  Now the
available android-cm-web will be used for these tasks instead.

The 'dmsutil.py' is updated to provide interface to get DMS title from
dms_tag_server using the new request, 'DMS_INFO'.

Change-Id: I5b59ab9871d000e9aa849c6d059797241b318d02