release-shortlog.py: Added new method 'dmsqueryShowTitle'
This change is to get rid of the dependency to delivery web tool for
generating DMS query file.

The delivery web tool was used to get the DMS title information and
to generate the DMS query file during the official build.  Now the
available android-cm-web will be used for these tasks instead.

The script 'dmsquery' in this git is still required to extract the
'FIX=DMSxxxxxxxx' pattern from the git log output.  This does not
involve the delivery web.

With the introduction of '--no-dw' option the 'release-shortlog.py'
is updated to use the DMS Tag Server instead of delivery web for
getting the DMS title information.  This option can be skipped if
delivery web needs to be used for legacy.

It also saves the list of DMS issues encountered in the commit
message to a text file with the '--log' option.

Change-Id: I62d2fbb7a734b40a07dcd3de0aece94170b428a9