manifestbranches.py: check for existence of default.xml
In the get_manifests method, `git show` is used to extract the
content of the manifest for the given refname.  However if the
default.xml file does not exist on that refname, an error will
occur.  Before attempting to extract default.xml, ensure that it
exists by calling `git ls-tree` and checking that the file is
listed.