run_query.pl: Added '-save' option
With the change in the release_shortlog.py script to use the
DMS Tag Server for generating the DMS information it is now required
to save the DMS query file generated by run_query.pl in a specific
name.  This will be helpful for uploading the generated query file
as a part of release information in the wiki page.

If a name is given the generated query files will be of the following
foramt:
<file-name>.qry
untagged_issues_<file-name>.qry
incorrect_tagged_issues_<file-name>.qry
placeholder_<file-name>.qry

Change-Id: I6520c62a0f86004cf764b1ad79c15426a57d7b45