gerrit.py: parameters to get_patchset_refspec must be integer
The change number and patch set number passed to get_patchset_refspec
must be integers, otherwise the % operation and string formatting with
%d will fail.

Added casts to int to force string parameters to integer.  This will
of course still fail if a string is passed that does not convert to a
valid integer, but at least prevents errors in case the caller passes
a valid integer as a string, i.e. "123".

Also tidied up a bit to fix errors reported by pylint.