dmsutil.py: limit number of DMS per request
To reduce the load on the tag server, the number of DMSs that are
sent per request is limited.  The list of DMSs will be split into
batches of 100, and sent to the server in multiple requests.

Additionally, the interface to the dms_with_title method is changed
from a comma-separated string to a Python list.

release-shortlog.py is changed to use the new interface.