Tag Server Fix: Send the error message back to client on failure
Update the tools chain to include the error information in the
message sent back to the client.

Also formatted the error message in the dmsutil.py to strip off
the 'SRV_ERROR' & 'SRV_END' strings from the message.

cherrypick.py is updated to include the branch name in the subject
line of the error mail.

run_query.pl is updated to throw an error message when it can't
establish a session with the DMS server and the same will be
propagated to the cherry pick script.

Change-Id: I976e7ca40876b6d0064d074bc30b11fb219e3abd