branch_policies.py: do not support regex in the branch name
Removed support for regular expression in the branch names.  We do
not use this feature; all branches in the policy file so far are
unique.

Furthermore, supporting regular expression potentially allows more
than one policy to apply to a given branch name.

Also updated the implementation to not allow the policy to include
more than one branch with the same name.

Change-Id: I815663d7677ef114a59b23d53291f3bb3061ae50