check_external_package: fix bug for not synced newly added packages.
As we check package existing via c2d web page to save time,
but we can't always rely on it, because the web page may not
be opened by some reason sometimes. Then, we should
check the package in c2d server, and if there's problem with c2d web
page for an existing package, we give a warning.

Change-Id: I62b584b190689c9380bd8eb8f64ba26a9e9507c2