release-shortlog.py: better error handling and other minor improvements
Use RepoXmlManifest to parse the manifest. This will reduce the chance
of errors occurring because path or revision is not in the element.

When an error occurs changing to a directory, just report the error
and continue with the next git.

Exit with an error if not enough arguments are given.

Define default parameters in the module scope, and update the default
job URL to something more recent than edream1.0.

Remove duplicates from the list of DMS issues.  Display the list of
issues sorted by DMS ID.

Fix a couple of other issues found by pylint.

Change-Id: I938bb2e12e7ca4b72b4895c4eaeec6ab273ac88d