cherrypick.py: Check for repo install before getting old cherries
Check that repo is installed before getting the list of old cherries
from the CM Server.  There is no point getting the list if repo is
not installed and the script cannot continue.

When the user does not pass working directory as an option, print
an info message that the current directory will be used.

Add the directory name when reporting that repo is not installed.

Also removed a superfluous comma in the manifest push operation.

Change-Id: I83de1604098442ef2c3d2f5d04d02da67f614be6