cherrypick.py: Replace --skip-review with --manifest-review
The --skip-review option causes the script to push manifest changes
directly to the repository, bypassing the Gerrit review.

The default is False, but it is always used and the user needs to
always add it on the command line.

This change replaces the option with --manifest-review, which when
set will cause the manifest change to be uploaded to Gerrit for
review.  The default is False, meaning changes will be pushed
directly to the repository.

With this change, the behaviour of the script will be the same,
but the user does not need to explicitly add an option on the
command line.

Change-Id: I08144915b30044915d6c8ec785ef527861bcdd00