cherry_update.py: Add option to switch between updating all or only new
Some of the cherrypick policies have a lot of cherrypicks with error
status that have not been cherrypicked with the -x option, and thus
will never be found by the update script.  Checking for these every
time the script is run takes a long time.

Add a new option "full".  Only when this option is set will all the
cherrypicks be checked for updates.  When not set (default) only the
ones with "new" status will be checked.

Based on this change, the usage of the script can be changed.  It
can be run much more frequently (maybe once per hour) for the new
changes, and less frequently (maybe daily) for the others.

Change-Id: I1d5711b47766807b58a070e8ff4d1f9584071102