cherry_update.py: update data for all active cherrypick policies
The current implementation of the script means that we need to set
up a separate Jenkins job to run it for each cherrypick policy.

Sometimes we forget to add the update job, so some cherrypick data
does not get updated.

Added the ability to download the list of manifest branches from
the CM server, and then for each branch that has a cherrypick
policy, update the cherrypick data.

Now we can remove the multiple manually configured update jobs in
Jenkins, and replace them with a single job that runs this script
and will automatically update any newly added policies.

It is still possible to run the update for a specific policy by
specifying the source, target, and manifest as command line
options.

Change-Id: I5b8daf44b4b4e5fa337167f500d6e7edda54b0c6