cm_server.py: return branch config as BranchPolicies object
In the get_branch_config method, instead of returning the branch
config as a raw XML string, which the calling script must use to
instantiate a BranchPolicies object, instantiate and return the
BranchPolicies object in the method.  This will reduce the amount
of repeated code in all the scripts that use this method.

Change-Id: I87e80c788e5547ca4a28eb71381338fa785b9236