cherrypick.py: check for cherrypick policy before checking for DMS tags
The cherrypick script checks for DMS tags on the given target branch
before checking that a cherrypick policy exists.  If the given target
branch does not exist in the list of branches, it will display an
error "must specify DMS tags" which does not make sense.

Instead, first check that a cherrypick policy exists.  Then if that
check passes, at least the DMS tag check is done against a target
branch that is known to exist.

Change-Id: I9e2072e9082597400c1e71c39f25b9f7b97de9f5